Integrated authentication

+++ base/views.py
@@ -1,10 +1,43 @@
 from django.shortcuts import render, redirect
+from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
+from django.contrib.auth.models import User
 from .forms import RoomForm
+from django.contrib.auth import authenticate,login,logout
+from django.http import HttpResponse
 # Create your views here.
 # Dump Business logic 
 
+
+def loginView(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist !!')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exist')
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+def logoutView(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     #implemeting the letter search
@@ -21,6 +54,11 @@ def room(request,pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+def registerView(request):
+    page = 'register'
+    return render(request,'base/login_register.html',page)
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -32,9 +70,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Do not have the rights !!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -45,9 +87,13 @@ def updateRoom(request,pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Do not have the rights")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
